requirements.txt
requests~=2.28.0
numpy~=1.22.4
matplotlib~=3.5.2
pillow~=9.1.1
beautifulsoup4~=4.11.1
selenium~=4.2.0
opencv-python~=4.6.0.66
scikit-image~=0.18.3
scikit-learn~=1.1.1